Added authentication + authorization and unit test

+++ src/main/java/com/vertxexploration/webapp/http/HttpServerVerticle.java
@@ -14,16 +14,35 @@ import com.vertxexploration.webapp.db.WikiDatabaseService;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
+import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.json.JsonObject;
+import io.vertx.core.net.JksOptions;
+import io.vertx.ext.auth.KeyStoreOptions;
+import io.vertx.ext.auth.jwt.JWTAuthOptions;
+import io.vertx.ext.auth.jwt.JWTOptions;
+import io.vertx.ext.auth.shiro.ShiroAuthOptions;
+import io.vertx.ext.auth.shiro.ShiroAuthRealmType;
 import io.vertx.ext.web.client.WebClientOptions;
 import io.vertx.reactivex.core.AbstractVerticle;
 import io.vertx.reactivex.core.http.HttpServer;
+import io.vertx.reactivex.ext.auth.AuthProvider;
+import io.vertx.reactivex.ext.auth.User;
+import io.vertx.reactivex.ext.auth.jwt.JWTAuth;
+import io.vertx.reactivex.ext.auth.shiro.ShiroAuth;
 import io.vertx.reactivex.ext.web.Router;
 import io.vertx.reactivex.ext.web.RoutingContext;
 import io.vertx.reactivex.ext.web.client.HttpResponse;
 import io.vertx.reactivex.ext.web.client.WebClient;
 import io.vertx.reactivex.ext.web.codec.BodyCodec;
+import io.vertx.reactivex.ext.web.handler.AuthHandler;
 import io.vertx.reactivex.ext.web.handler.BodyHandler;
+import io.vertx.reactivex.ext.web.handler.CookieHandler;
+import io.vertx.reactivex.ext.web.handler.FormLoginHandler;
+import io.vertx.reactivex.ext.web.handler.JWTAuthHandler;
+import io.vertx.reactivex.ext.web.handler.RedirectAuthHandler;
+import io.vertx.reactivex.ext.web.handler.SessionHandler;
+import io.vertx.reactivex.ext.web.handler.UserSessionHandler;
+import io.vertx.reactivex.ext.web.sstore.LocalSessionStore;
 import io.vertx.reactivex.ext.web.templ.FreeMarkerTemplateEngine;
 
 public class HttpServerVerticle extends AbstractVerticle {
@@ -39,6 +58,8 @@ public class HttpServerVerticle extends AbstractVerticle {
 	// For auto redeployment for code changes, put --redeploy="src/**/*.java"
 	// --launcher-class=io.vertx.core.Launcher to Program argument
 	private WikiDatabaseService dbService;
+	private AuthProvider auth;
+	private JWTAuth jwtAuth;
 
 	@Override
 	public void start(Future<Void> startFuture) throws Exception {
@@ -46,18 +67,63 @@ public class HttpServerVerticle extends AbstractVerticle {
 		String wikiDbQueue = config().getString(CONFIG_WIKIDB_QUEUE, "wikidb.queue");
 		dbService = WikiDatabaseService.createProxy(vertx.getDelegate(), wikiDbQueue);
 
-		HttpServer server = vertx.createHttpServer();
+		this.auth = ShiroAuth.create(vertx, new ShiroAuthOptions()
+				.setType(ShiroAuthRealmType.PROPERTIES)
+				.setConfig(new JsonObject()
+						.put("properties_path", "classpath:wiki-users.properties")));
+		
+		HttpServer server = vertx.createHttpServer(new HttpServerOptions()
+			.setSsl(true)
+			.setKeyStoreOptions(new JksOptions()
+				.setPath("server-keystore.jks")
+				.setPassword("secret")));
 
 		Router router = Router.router(vertx);
+		
+		// Security config
+		router.route().handler(CookieHandler.create());
+		router.route().handler(BodyHandler.create());
+		router.route().handler(SessionHandler.create(LocalSessionStore.create(vertx)));
+		router.route().handler(UserSessionHandler.create(auth));  
+
+		AuthHandler authHandler = RedirectAuthHandler.create(auth, "/login"); 
+		router.route("/").handler(authHandler);  
+		router.route("/wiki/*").handler(authHandler);
+		router.route("/action/*").handler(authHandler);
+		
+		// login/logout
+		router.get("/login").handler(this::loginHandler);
+		router.post("/login-auth").handler(FormLoginHandler.create(auth));  
+		router.get("/logout").handler(context -> {
+		  context.clearUser();  
+		  context.response()
+		    .setStatusCode(302)
+		    .putHeader("Location", "/")
+		    .end();
+		});
+		
+		// Website gateway
 		router.get("/").handler(this::indexHandler);
 		router.get("/wiki/:page").handler(this::pageRenderingHandler);
 		router.post().handler(BodyHandler.create());
-		router.post("/save").handler(this::pageUpdateHandler);
-		router.post("/create").handler(this::pageCreateHandler);
-		router.post("/delete").handler(this::pageDeletionHandler);
-		router.get("/backup").handler(this::backupHandler);
+		router.post("/action/save").handler(this::pageUpdateHandler);
+		router.post("/action/create").handler(this::pageCreateHandler);
+		router.post("/action/delete").handler(this::pageDeletionHandler);
+		router.get("/action/backup").handler(this::backupHandler);
 
+		
+		// API Gateway
 		Router apiRouter = Router.router(vertx);
+		
+
+		this.jwtAuth = JWTAuth.create(vertx, 
+			new JWTAuthOptions().setKeyStore(
+				new KeyStoreOptions()
+				.setPath("keystore.jceks")
+				.setPassword("secret")));
+
+		apiRouter.route().handler(JWTAuthHandler.create(jwtAuth, "/api/token"));
+		
 		apiRouter.get("/pages").handler(this::apiRoot);
 		apiRouter.get("/pages/:id").handler(this::apiGetPage);
 		apiRouter.post().handler(BodyHandler.create());
@@ -65,6 +131,8 @@ public class HttpServerVerticle extends AbstractVerticle {
 		apiRouter.put().handler(BodyHandler.create());
 		apiRouter.put("/pages/:id").handler(this::apiUpdatePage);
 		apiRouter.delete("/pages/:id").handler(this::apiDeletePage);
+		apiRouter.get("/token").handler(this::jwtTokenGeneratorHandler);
+		
 		router.mountSubRouter("/api", apiRouter);
 
 		int portNumber = config().getInteger(CONFIG_HTTP_SERVER_PORT, 8080);
@@ -78,55 +146,114 @@ public class HttpServerVerticle extends AbstractVerticle {
 			}
 		});
 	}
+	
+	private void jwtTokenGeneratorHandler(RoutingContext context) {
+		JsonObject creds = new JsonObject()
+				.put("username", context.request().getHeader("login"))
+				.put("password", context.request().getHeader("password"));
+		
+		auth.authenticate(creds, authResult -> { // this is to authenticate the user id
 
-	private final FreeMarkerTemplateEngine templateEngine = FreeMarkerTemplateEngine.create();
+			if (authResult.succeeded()) {
+				
+				// From here above, it is to authorize the user based on wiki-user.properties
+				User user = authResult.result();
+				user.isAuthorised("create", canCreate -> {
+					user.isAuthorised("delete", canDelete -> {
+						user.isAuthorised("update", canUpdate -> {
 
-	private void indexHandler(RoutingContext context) {
-		dbService.fetchAllPages(reply -> {
-			if (reply.succeeded()) {
-				context.put("title", "Wiki home");
-				context.put("pages", reply.result().getList());
-				templateEngine.render(context, "templates", "/index.ftl", ar -> {
-					if (ar.succeeded()) {
-						context.response().putHeader("Content-Type", "text/html");
-						context.response().end(ar.result());
-					} else {
-						context.fail(ar.cause());
-					}
+							String token = this.jwtAuth.generateToken(
+									new JsonObject().put("username", context.request().getHeader("login"))
+											.put("canCreate", canCreate.succeeded() && canCreate.result())
+											.put("canDelete", canDelete.succeeded() && canDelete.result())
+											.put("canUpdate", canUpdate.succeeded() && canUpdate.result()),
+									new JWTOptions().setSubject("Wiki API").setIssuer("Vert.x"));
+							context.response().putHeader("Content-Type", "text/plain").end(token);
+						});
+					});
 				});
 			} else {
-				context.fail(reply.cause());
+				context.fail(401);
+			}
+		});
+	}
+	
+	private void loginHandler(RoutingContext context) {
+		context.put("title", "Login");
+		templateEngine.render(context, "templates", "/login.ftl", ar -> {
+			if (ar.succeeded()) {
+				context.response().putHeader("Content-Type", "text/html");
+				context.response().end(ar.result());
+			} else {
+				context.fail(ar.cause());
 			}
 		});
 	}
 
+	private final FreeMarkerTemplateEngine templateEngine = FreeMarkerTemplateEngine.create();
+
+	private void indexHandler(RoutingContext context) {
+		context.user().isAuthorised("create", res -> {
+			boolean canCreatePage = res.succeeded() && res.result();
+			dbService.fetchAllPages(reply -> {
+				if (reply.succeeded()) {
+					context.put("title", "Wiki home");
+					context.put("pages", reply.result().getList());
+					context.put("canCreatePage", canCreatePage); // to be used in the index.ftl template
+					context.put("username", context.user().principal().getString("username"));
+					templateEngine.render(context, "templates", "/index.ftl", ar -> {
+						if (ar.succeeded()) {
+							context.response().putHeader("Content-Type", "text/html");
+							context.response().end(ar.result());
+						} else {
+							context.fail(ar.cause());
+						}
+					});
+				} else {
+					context.fail(reply.cause());
+				}
+			});
+		});
+	}
+
 	private void pageRenderingHandler(RoutingContext context) {
-		String requestedPage = context.request().getParam("page");
-		dbService.fetchPage(requestedPage, reply -> {
-			if (reply.succeeded()) {
+		context.user().isAuthorized("update", updateResponse -> {
+			boolean canSavePage = updateResponse.succeeded() && updateResponse.result();
+			context.user().isAuthorized("delete", deleteResponse -> {
+				boolean canDeletePage = deleteResponse.succeeded() && deleteResponse.result();
 
-				JsonObject payLoad = reply.result();
-				boolean found = payLoad.getBoolean("found");
-				String rawContent = payLoad.getString("rawContent", EMPTY_PAGE_MARKDOWN);
-				context.put("title", requestedPage);
-				context.put("id", payLoad.getInteger("id", -1));
-				context.put("newPage", found ? "no" : "yes");
-				context.put("rawContent", rawContent);
-				context.put("content", Processor.process(rawContent));
-				context.put("timestamp", new Date().toString());
+				String requestedPage = context.request().getParam("page");
+				dbService.fetchPage(requestedPage, reply -> {
+					if (reply.succeeded()) {
 
-				templateEngine.render(context, "templates", "/page.ftl", ar -> {
-					if (ar.succeeded()) {
-						context.response().putHeader("Content-Type", "text/html");
-						context.response().end(ar.result());
+						JsonObject payLoad = reply.result();
+						boolean found = payLoad.getBoolean("found");
+						String rawContent = payLoad.getString("rawContent", EMPTY_PAGE_MARKDOWN);
+						context.put("title", requestedPage);
+						context.put("id", payLoad.getInteger("id", -1));
+						context.put("newPage", found ? "no" : "yes");
+						context.put("rawContent", rawContent);
+						context.put("content", Processor.process(rawContent));
+						context.put("timestamp", new Date().toString());
+						context.put("username", context.user().principal().getString("username"));
+						context.put("canSavePage", canSavePage);
+						context.put("canDeletePage", canDeletePage);
+
+						templateEngine.render(context, "templates", "/page.ftl", ar -> {
+							if (ar.succeeded()) {
+								context.response().putHeader("Content-Type", "text/html");
+								context.response().end(ar.result());
+							} else {
+								context.fail(ar.cause());
+							}
+						});
+
 					} else {
-						context.fail(ar.cause());
+						context.fail(reply.cause());
 					}
 				});
 
-			} else {
-				context.fail(reply.cause());
-			}
+			});
 		});
 	}
 
@@ -145,9 +272,21 @@ public class HttpServerVerticle extends AbstractVerticle {
 
 		String markdown = context.request().getParam("markdown");
 		if ("yes".equals(context.request().getParam("newPage"))) {
-			dbService.createPage(title, markdown, handler);
+			context.user().isAuthorised("create", res -> {
+				if (res.succeeded() && res.result()) {
+					dbService.createPage(title, markdown, handler);
+				} else {
+					context.response().setStatusCode(403).end();
+				}
+			});
 		} else {
-			dbService.savePage(Integer.valueOf(context.request().getParam("id")), markdown, handler);
+			context.user().isAuthorised("update", res -> {
+				if (res.succeeded() && res.result()) {
+					dbService.savePage(Integer.valueOf(context.request().getParam("id")), markdown, handler);
+				} else {
+					context.response().setStatusCode(403).end();
+				} 
+			});
 		}
 	}
 
@@ -163,13 +302,22 @@ public class HttpServerVerticle extends AbstractVerticle {
 	}
 
 	private void pageDeletionHandler(RoutingContext context) {
-		dbService.deletePage(Integer.valueOf(context.request().getParam("id")), reply -> {
-			if (reply.succeeded()) {
-				context.response().setStatusCode(303);
-				context.response().putHeader("Location", "/");
-				context.response().end();
+		context.user().isAuthorised("delete", res -> {
+			if (res.succeeded() && res.result()) {
+
+				// Original code:
+				dbService.deletePage(Integer.valueOf(context.request().getParam("id")), reply -> {
+					if (reply.succeeded()) {
+						context.response().setStatusCode(303);
+						context.response().putHeader("Location", "/");
+						context.response().end();
+					} else {
+						context.fail(reply.cause());
+					}
+				});
+
 			} else {
-				context.fail(reply.cause());
+				context.response().setStatusCode(403).end();
 			}
 		});
 	}
@@ -192,35 +340,30 @@ public class HttpServerVerticle extends AbstractVerticle {
 				});
 
 				webClient.post(443, "api.github.com", "/gists").putHeader("Accept", "application/vnd.github.v3+json")
-						.putHeader("Content-Type", "application/json").as(BodyCodec.jsonObject())
-						.sendJsonObject(gistPayload, ar -> { // This code will
-																// call all the
-																// way down
-																// until netty
-																// server
-																// implementation
-							if (ar.succeeded()) {
-								HttpResponse<JsonObject> response = ar.result();
-								if (response.statusCode() == 201) {
-									context.put("backup_gist_url", response.body().getString("html_url"));
-									indexHandler(context);
-								} else {
-									StringBuilder message = new StringBuilder().append("Could not backup the wiki: ")
-											.append(response.statusMessage());
-									JsonObject body = response.body();
-									if (body != null) {
-										message.append(System.getProperty("line.separator"))
-												.append(body.encodePrettily());
-									}
-									LOGGER.error(message.toString());
-									context.fail(502);
-								}
+					.putHeader("Content-Type", "application/json").as(BodyCodec.jsonObject())
+					.sendJsonObject(gistPayload, ar -> { // This code will call all the way down until netty server implementation
+						if (ar.succeeded()) {
+							HttpResponse<JsonObject> response = ar.result();
+							if (response.statusCode() == 201) {
+								context.put("backup_gist_url", response.body().getString("html_url"));
+								indexHandler(context);
 							} else {
-								Throwable err = ar.cause();
-								LOGGER.error("HTTP Client error", err);
-								context.fail(err);
+								StringBuilder message = new StringBuilder().append("Could not backup the wiki: ")
+										.append(response.statusMessage());
+								JsonObject body = response.body();
+								if (body != null) {
+									message.append(System.getProperty("line.separator"))
+											.append(body.encodePrettily());
+								}
+								LOGGER.error(message.toString());
+								context.fail(502);
 							}
-						});
+						} else {
+							Throwable err = ar.cause();
+							LOGGER.error("HTTP Client error", err);
+							context.fail(err);
+						}
+					});
 
 			} else {
 				context.fail(reply.cause());
@@ -350,9 +493,13 @@ public class HttpServerVerticle extends AbstractVerticle {
 	}
 	
 	private void apiDeletePage(RoutingContext context) {
-		int id = Integer.valueOf(context.request().getParam("id"));
-		dbService.deletePage(id, reply -> {
-			handleSimpleDbReply(context, reply);
-		});
+		if (context.user().principal().getBoolean("canDelete", false)) {
+			int id = Integer.valueOf(context.request().getParam("id"));
+			dbService.deletePage(id, reply -> {
+				handleSimpleDbReply(context, reply);
+			});
+		} else {
+			context.fail(401);
+		}
 	}
 }
